Prevent canvas always occupying entire viewport.
Usually when used as a background with a <div>, it's common to set
position:absolute so that the particles occupies the container div.
However, the canvas will be set to the absolute div's size using JS,
which will occupy the entire viewport, even if the container div does
not.

This fixes that by using the size of the container's parent div which is
not position:absolute.